Authzen: clarify performApiRequestCall single-200 check + body-extraction gate
Two cosmetic readability fixes in AbstractAuthzenPDPTest:

performApiRequestCall (line 316): the singleton-200 check used
`acceptable.iterator().next() == 200`, which reads as reference
equality at a glance even though autoboxing makes it numerically
correct. Make the int-unbox explicit via .intValue() so the comparison
is unambiguously primitive.

performAuthzenApiFlow (line 198): add an in-code note that the
body-extraction gate fires whenever 200 is in the acceptable set, so a
future test author who introduces a mixed Set.of(200, 4xx) acceptance
knows the response-shape validators will run unconditionally on
non-2xx bodies under that configuration.

No behavior change.

+++ src/main/java/net/openid/conformance/authzen/AbstractAuthzenPDPTest.java
@@ -195,6 +195,11 @@ public abstract class AbstractAuthzenPDPTest extends AbstractRedirectServerTestM
 		eventLog.startBlock("Make request to API endpoint");
 		createAuthzenApiRequest();
 		performSingleApiRequest();
+		// Body-extraction / response validation runs whenever 200 is in the
+		// acceptable set. For tests that allow a mix of 200 and non-200
+		// codes, this runs unconditionally — author must ensure
+		// processAuthApiEndpointResponse() / validateAuthApiEndpointResponse()
+		// tolerate a 4xx body shape if such a mixed set is ever used.
 		if (getAcceptableHttpStatusCodes().contains(200)) {
 			processAuthApiEndpointResponse();
 			validateAuthApiEndpointResponse();
@@ -308,7 +313,7 @@ public abstract class AbstractAuthzenPDPTest extends AbstractRedirectServerTestM
 
 	protected void performApiRequestCall() {
 		Set<Integer> acceptable = getAcceptableHttpStatusCodes();
-		if (acceptable.size() == 1 && acceptable.iterator().next() == 200) {
+		if (acceptable.size() == 1 && acceptable.iterator().next().intValue() == 200) {
 			call(sequence(CallAuthzenApiEndpointAndVerifySuccessfulResponse.class));
 		} else {
 			JsonObject wrapper = new JsonObject();
